abstract out parseDeclarations
