+++ src/flowCoffeescriptKernel/flowCoffeescriptKernel.js
@@ -1,7 +1,7 @@
 import { safetyWrapCoffeescript } from './safetyWrapCoffeescript';
 import { compileCoffeescript } from './compileCoffeescript';
 import { parseJavascript } from './parseJavascript';
-import { identifyDeclarations } from './identifyDeclarations';
+import { parseDeclarations } from './parseDeclarations';
 
 export function flowCoffeescriptKernel() {
   const lodash = window._;
@@ -9,28 +9,6 @@ export function flowCoffeescriptKernel() {
   const escodegen = window.escodegen;
   const esprima = window.esprima;
   const CoffeeScript = window.CoffeeScript;
-  const parseDeclarations = block => {
-    let declaration;
-    let declarations;
-    let node;
-    let _i;
-    let _j;
-    let _len;
-    let _len1;
-    const identifiers = [];
-    const _ref = block.body;
-    for (_i = 0, _len = _ref.length; _i < _len; _i++) {
-      node = _ref[_i];
-      declarations = identifyDeclarations(node);
-      if (declarations) {
-        for (_j = 0, _len1 = declarations.length; _j < _len1; _j++) {
-          declaration = declarations[_j];
-          identifiers.push(declaration);
-        }
-      }
-    }
-    return lodash.indexBy(identifiers, identifier => identifier.name);
-  };
   const traverseJavascript = (parent, key, node, f) => {
     let child;
     let i;
